+++ src/main/java/com/example/buensaborback/domain/entities/Empresa.java
@@ -17,6 +17,6 @@ public class Empresa extends Base{
 
     private String nombre;
     private String razonSocial;
-    private Integer cuil;
+    private Long cuil;
 
 }
